Home frag save video palying state, disable back from home, finish, restart button bg color disable change, edit text set cursor to end trial, fix strings.

+++ RandomMemories/app/src/main/java/com/example/randommemories/ui/main/LinedEditText.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.graphics.*
 import android.os.Build
 import android.util.AttributeSet
-import android.widget.EditText
 import androidx.annotation.RequiresApi
 import com.example.randommemories.R
 
@@ -13,6 +12,7 @@ import com.example.randommemories.R
 class LinedEditText(context: Context, attrs: AttributeSet?) : androidx.appcompat.widget.AppCompatEditText(context, attrs) {
     private val mRect: Rect = Rect()
     private val mPaint: Paint = Paint()
+//    private var isFirstClick = true
 
     // we need this constructor for LayoutInflater
     init {
@@ -31,10 +31,19 @@ class LinedEditText(context: Context, attrs: AttributeSet?) : androidx.appcompat
         val paint: Paint = mPaint
         var baseline = getLineBounds(0, r)+6f //first line
         for (i in 0 until count) {
-            canvas.drawLine(r.left.toFloat(), (baseline + 1).toFloat(),
-                r.right.toFloat(), (baseline + 1).toFloat(), paint)
+            canvas.drawLine(r.left.toFloat(), (baseline + 1),
+                r.right.toFloat(), (baseline + 1), paint)
             baseline += (lineHeight+0.2f) //next line
         }
         super.onDraw(canvas)
     }
+
+//    override fun onSelectionChanged(selStart: Int, selEnd: Int) {
+//        super.onSelectionChanged(selStart, selEnd)
+//        if (isFirstClick) {
+//            isFirstClick = false
+//            // Move cursor to the end of the text
+//            setSelection(text?.length ?: 0)
+//        }
+//    }
 }
